+++ src/main/scala/com/ucesys/sparklens/helper/HDFSConfigHelper.scala
@@ -19,16 +19,10 @@ package com.ucesys.sparklens.helper
 
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.SparkConf
-import org.apache.spark.deploy.SparkHadoopUtil
 
 object HDFSConfigHelper {
 
    def getHadoopConf(sparkConfOptional:Option[SparkConf]): Configuration = {
-    if (sparkConfOptional.isDefined) {
-      SparkHadoopUtil.get.newConfiguration(sparkConfOptional.get)
-    }else {
-      val sparkConf = new SparkConf()
-      SparkHadoopUtil.get.newConfiguration(sparkConf)
-    }
+       new Configuration()
   }
 }