+++ src/main/scala/com/ucesys/sparkscope/io/CsvHadoopMetricsLoader.scala
@@ -1,10 +1,13 @@
 package com.ucesys.sparkscope.io
 
 import com.ucesys.sparklens.common.AppContext
+import com.ucesys.sparklens.timespan.ExecutorTimeSpan
 import com.ucesys.sparkscope.SparkScopeAnalyzer.{DriverCsvMetrics, ExecutorCsvMetrics}
 import com.ucesys.sparkscope.SparkScopeConfig
 import com.ucesys.sparkscope.data.DataFrame
 import com.ucesys.sparkscope.utils.Logger
+
+import scala.collection.mutable
 
 class CsvHadoopMetricsLoader(readerFactory: FileReaderFactory,
                              appContext: AppContext,
@@ -14,13 +17,13 @@ class CsvHadoopMetricsLoader(readerFactory: FileReaderFactory,
   def load(): DriverExecutorMetrics = {
     val ac = appContext.filterByStartAndEndTime(appContext.appInfo.startTime, appContext.appInfo.endTime)
 
-    log.println(s"Reading driver metrics from ${sparkScopeConf.driverMetricsDir}, executor metrics from ${sparkScopeConf.executorMetricsDir}")
+    log.info(s"Reading driver metrics from ${sparkScopeConf.driverMetricsDir}, executor metrics from ${sparkScopeConf.executorMetricsDir}")
 
-    log.println("Reading driver metrics...")
+    log.info("Reading driver metrics...")
     val driverMetrics: Seq[DataFrame] = DriverCsvMetrics.map { metric =>
       val metricsFilePath = s"${sparkScopeConf.driverMetricsDir}/${appContext.appInfo.applicationID}.driver.${metric}.csv"
       val csvFileStr = readerFactory.getFileReader(metricsFilePath).read(metricsFilePath).replace("value", metric)
-      log.println(s"Reading ${metric} metric for driver from " + metricsFilePath)
+      log.info(s"Reading ${metric} metric for driver from " + metricsFilePath)
       DataFrame.fromCsv(metric, csvFileStr, ",")
     }
 
@@ -41,11 +44,16 @@ class CsvHadoopMetricsLoader(readerFactory: FileReaderFactory,
         metric.dropNLastRows(metric.numRows-driverMetricsAlignedSize)
     }
 
-    log.println("Reading executor metrics...")
-    val executorsMetricsMap: Map[Int, Seq[DataFrame]] = ac.executorMap.map { case (executorId, _) =>
+    log.info("Reading executor metrics...")
+
+    // Filter out executorId="driver" which occurs in local mode
+    val executorsMetricsMapNonDriver: mutable.HashMap[String, ExecutorTimeSpan] = ac.executorMap
+        .filter {case (executorId, _) => executorId != "driver"}
+
+    val executorsMetricsMap: Map[String, Seq[DataFrame]] = executorsMetricsMapNonDriver.map { case (executorId, _) =>
       val metricTables: Seq[DataFrame] = ExecutorCsvMetrics.flatMap { metric =>
         val metricsFilePath = s"${sparkScopeConf.executorMetricsDir}/${appContext.appInfo.applicationID}.${executorId}.${metric}.csv"
-        log.println(s"Reading ${metric} metric for executor=${executorId} from " + metricsFilePath)
+        log.info(s"Reading ${metric} metric for executor=${executorId} from " + metricsFilePath)
         val csvFileStrOpt = try {
           Some(readerFactory.getFileReader(metricsFilePath).read(metricsFilePath).replace("value", metric).replace("count", metric))
         }
@@ -58,11 +66,14 @@ class CsvHadoopMetricsLoader(readerFactory: FileReaderFactory,
       }
 
       metricTables match {
-        case seq: Seq[DataFrame] if (seq.length == ExecutorCsvMetrics.length) => (executorId.toInt, Some(metricTables))
-        case _ => log.warn(s"Missing metrics for executor=${executorId}");(executorId.toInt, None)
+        case seq: Seq[DataFrame] if (seq.length == ExecutorCsvMetrics.length) => (executorId, Some(metricTables))
+        case _ => log.warn(s"Missing metrics for executor=${executorId}");(executorId, None)
       }
-
     }.retain{case(_, opt) => opt.nonEmpty}.map{case(id, opt) => (id, opt.get)}.toMap
+
+    if(executorsMetricsMap.isEmpty) {
+      throw new IllegalArgumentException("No executor metrics found")
+    }
 
     val executorsMetricsMapAligned = executorsMetricsMap.map{case (executorId, metrics) => {
       val executorMetricsAlignedSize = metrics.map(_.numRows).min