+++ src/test/scala/com/ucesys/sparkscope/io/CsvHadoopMetricsLoaderSuite.scala
@@ -20,13 +20,9 @@ package com.ucesys.sparkscope.io
 
 import com.ucesys.sparkscope.SparkScopeAnalyzer.{DriverCsvMetrics, ExecutorCsvMetrics}
 import com.ucesys.sparkscope.TestHelpers._
-import com.ucesys.sparkscope.io.{CsvHadoopMetricsLoader, HadoopFileReader, HadoopPropertiesLoader, PropertiesLoader, PropertiesLoaderFactory}
 import org.apache.spark.SparkConf
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.GivenWhenThen
-
-import java.io.FileNotFoundException
-import java.util.Properties
 import org.scalatest.FunSuite
 
 class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWhenThen {
@@ -41,7 +37,7 @@ class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWh
 
     Then("Driver and Executor Metrics should be loaded")
     assert(driverExecutorMetrics.driverMetrics.length == DriverCsvMetrics.length)
-    assert(driverExecutorMetrics.executorMetricsMap(1).length == ExecutorCsvMetrics.length)
+    assert(driverExecutorMetrics.executorMetricsMap("1").length == ExecutorCsvMetrics.length)
 
     And("CsvHadoopMetricsLoader should ignore extra rows for driver metrics")
     driverExecutorMetrics.driverMetrics.foreach{ metric =>
@@ -49,7 +45,7 @@ class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWh
     }
 
     And("CsvHadoopMetricsLoader should ignore extra rows for executor metrics")
-    driverExecutorMetrics.executorMetricsMap(1).foreach { metric =>
+    driverExecutorMetrics.executorMetricsMap("1").foreach { metric =>
       assert(metric.numRows == 10)
     }
   }
@@ -91,6 +87,6 @@ class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWh
     assert(driverExecutorMetrics.executorMetricsMap.head._2.length == ExecutorCsvMetrics.length)
 
     And("Missing Executor Metrics should be skipped")
-    assert(driverExecutorMetrics.executorMetricsMap.get(6).isEmpty)
+    assert(driverExecutorMetrics.executorMetricsMap.get("6").isEmpty)
   }
 }