+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
+
+## [0.1.2] - 2023-10-10
+- Rename package from sparkscope to sparkscope-spark2
+- Bump Scala version from 2.11.8 to 2.11.12
+- Handle gently case for master=local where executorId=driver
+
 ## [0.1.1] - 2023-10-06
-Initial SparkScope release:
 - configuration:
   - sparkscope will prioritize csv metrics SparkConf configuration and use metrics.properties files as last-resort 
 - html-report: